source/requirements.txt
flask==0.11
flask-debug==0.4.3
flask-debugtoolbar==0.10.0
flask-marshmallow==0.7.0
flask-restful==0.3.5
flask-sqlalchemy==2.1
marshmallow-sqlalchemy==0.12.0
uwsgi==2.0.13.1
